Extract list-item view models and split subtitle into public sub-views
Move ChatChannelListItemViewModel and ChatThreadListItemViewModel to
dedicated files, and break the channel list item subtitle area into
small, primitive-driven public views (title, muted icon, failed-to-send,
draft/deleted/author/attachment previews) that customers can compose
or replace independently.
